auto-fix-qortal script for automated core update
This script will check the hash of the release version of Qortal (in a way that will surely be modified in the future) - It will then verify that the installed qortal.jar is the same hash as the new release. If so, it will exit.

If the installed qortal.jar is a DIFFERENT hash than release version, it will remove existing version, replace with new version, remove db, log files, update start script to use a version that uses max RAM for the JVM of 2.2GB, then start Qortal so that it will auto-bootstrap.

This is a VERY early and VERY 'rough' version of the script, but it works.

It needs a lot more checks to ensure it doesn't remove qortal with nothing to replace it, etc...

it NEEDS internet connectivity, and checks for that.

Many other additions will happen as well.