simple script to stop/start qortal
This script is build to be placed on desktop for linux machines, and will stop/start qortal core for them. This can be used for any linux machine after running auto-fix script from cron and not showing qortal icon, if the user wants to see the icon.

it can also be used by QORTector to start Qortal in a way that shows the icon easily.